+++ apps/web/src/components/modal.tsx
@@ -8,7 +8,7 @@ type ModalProps = React.PropsWithChildren & {
 
 export function Modal({ children, title, onClose }: ModalProps) {
   return (
-    <div className='fixed top-0 left-0 flex w-full h-full z-10 bg-[#0004] backdrop-blur-lg items-center justify-center py-4 xs:px-4 sm:px-32 lg:px-64 xl:px-96'>
+    <div className='fixed top-0 left-0 flex w-full h-full z-90 bg-[#0004] backdrop-blur-lg items-center justify-center py-4 xs:px-4 sm:px-32 lg:px-64 xl:px-96'>
       <div className='bg-white rounded-xl shadow-lg p-4 w-full animate-slide-up'>
         <div className='flex items-center w-full justify-between mb-4'>
           <h3 className='font-semibold'>{title}</h3>